requirements.txt
streamlit
torch
pandas
sentence-transformers
pdfminer.six
